refactor: change strucrure of render
