+++ src/app.ts
@@ -1,7 +1,7 @@
 import './scss/styles.scss';
 import PageController from './pages/PageController';
 
-import './pages/home-page/staticElements/style.scss';
+import './pages/staticElements/style.scss';
 import './pages/home-page/postElements/style.scss'
 import Router from './router';
 import { sessionId } from './types/constants';
