+++ src/pages/home-page/HomePageController.ts
@@ -1,6 +1,6 @@
 import {PostElementsController, page} from './postElements/postElementsController';
-import HeaderController from './staticElements/HeaderController';
-import HeaderView from './staticElements/HeaderView';
+import HeaderController from '../staticElements/HeaderController';
+import HeaderView from '../staticElements/HeaderView';
 
 class HomePageController {
     HeaderController: HeaderController;
