+++ src/pages/staticElements/style.scss
